[WASH-390] TozID login for iOS
- Adds login method for iOS, which uses the updated e3db-swift and returns a Client
- Converts E3db Config to FlutterConfig so dart can decode the json.
- Updates permissions to allow navigation to http addresses.

+++ lib/tozny_model.dart
@@ -219,7 +219,7 @@ class AgentToken {
   @JsonKey(name: "token_type")
   String tokenType;
   @JsonKey(name: "expiry")
-  int expiry;
+  double expiry;
 
   AgentToken(
     this.accessToken,
